Fix a bug in Kruijswijk's resolve function
This bug was only encountered when adjudicating Pandin's paradox (DATC
6.F.16). There likely is a better fix for this problem, but this works
for now. Unluckly this removes the upper bound of MAX_ORDERS to
resolve's dependency list; it is not immediately clear how large this
list can grow now, but tests have shown it to remain tipically under
8 when running the DATC test suite. We can probably get away for now
with a fixed maximum size of 256, but further investigation is needed
to avoid the risk of buffer overflows.