Add "exclude" parameter to is_legal_move
When considering the legality of a move from a support order's
standpoint, there are times in which the move could only be legal, if
and only if the supporting unit were to have a convoy order for that
same move. This is of course impossible, and although the moving unit
cannot know about it, and therefore is legal when considered in its
own, the support shold be aware of it, and consider the move illegal
accordingly. Therefore, the exclude "parameter" allows to specify the
supporting territory to exclude in the search for a possible convoy
path.