Fix retreats issues with split coast territories
This was an old bug: the code would ignore coasts during retreats
computation, and it would therefore offer, for example, a retreat from
Spa to Mar, even when the unit retreating was a fleet stationed along
Spa's northern coast. Furthermore, when designating as an available
destination for a retreat a territory that had more than one coast, it
wouldn't tell which of the two it was actually possible to reach and
retreat to, if not both.
All of this has now been fixed.